Optimize tag matching
Instead of calling mystrcmp for each tag (which tries to match the tag
with strncmp), introduce a state machine (corresponds to a DFA) to match
tags.

This is pretty naive, and uses a full-sized (256 entries) table for each
state, requiring a total of 51 states for 45 tokens, amounting to 30KB
of state tables.