Do font supersampling the VSFilter way
To avoid jerking when font size scale is animated caused by grid
fitting, we intentionally request an overly large font size from
FreeType and downscale the outline afterwards.

Currently this large font size is fixed at 256.
This commit changes it to 64 times the actual wanted font size,
which is what is used by VSFilter. The advantages are:
  * Compatibility with VSFilter.
  * Support for font hinting, even if at 64 times the size.
    Knowing this coefficient, scripts can compensate
    for it by correspondingly decreasing \fs values.

This improves upon commits 7af7801 and c207000.

The drawback is that jerking is reintroduced when an mplayer
window with libass EOSD is resized. Jerking could be avoided
by scaling \fscx and \fscy instead of \fs. Pro: this would make
the exact look of the font completely resolution-independent.
Con: this would play badly with font hinting.
In any case, this commit makes no attempt to do this.

ensure_font_size, unused since 7af7801, is removed in favor
of clipping the font size just before passing it to FreeType,
as discussed in Google Code issue #46.
The mysterious height*2 check is dropped.

The scale_x/y check in get_outline_glyph used double_to_d6,
which now suffers from underflow. Bring it in line with
fill_glyph_hash and use double_to_d16 instead.