Use actual NBSP (\h) glyphs and always fall back to Arial
\h is often used in typesetting purely for horizontal padding.
This relies on its metrics being known. VSFilter uses the NBSP
glyph directly and substitutes Arial if it does not exist in
the given font. This commit adds the same behavior to libass.

This does not make our font selection closer to VSFilter's
in the general case. However, \h is particularly common
and sensitive, so this commit fixes it especially.

Fixes libass#2.
Fixes libass#42.