Close gaps in borders of self-intersecting glyphs
If the glyph is self-intersecting, so are the stroke
borders produced by stroking it. Normally, there is an
"inside" and an "outside" stroke border, but they swap
roles around self-intersections, the "inside" one going
on the outside while the "outside" one stays inside.
Currently, we take only the "outside" border, which thus
hides behind the main glyph around self-intersections.

Keep both stroke borders instead.

NOPE: this gives only a thick path around the edges
of the glyph, which breaks blur.