fontselect: replace is_postscript flag with check_postscript function
DirectWrite does not provide fast access to the is_postscript flag,
requiring each font to be loaded before its format can be determined.
Eagerly doing this for every installed font can be quite slow,
on the order of seconds.

To improve performance, ask the font provider for this information
only when it is actually needed, i.e. when one of the font's full
names or its PostScript name matches a requested font name and we
need to know whether to accept this match.

This makes font providers other than DirectWrite slightly slower.
This should not be a problem, but if it is, is_postscript can be
cached in their font private data.