Add meson as a secondary build system
This will bring back a first-party buildsystem for MSVC
after the previous solution https://github.com/libass/libass-msvc
was retired and make it easier to use libass as a subproject
in other meson projects.

It does not currently support all of our utilities nor testing
and fuzzing infrastructure and x86_32 assembly is always built in
PIC mode.
Due to meson deficiencies shared libraries lack symbol export control
and it is not possible to build x32 assembly.

However, for packaging and downstream use it is feature complete on
Windows and static-only builds are sufficient for subproject use on any
platform, so it brings a net positive and keeping this level of support
up alongside autotools is expected to be barely any effort.
And ofc future meson versions and patches
to this build setup might improve it further.

The port was initially brought up by Ryan Lucia (CoffeeFlux)
and later further adapted and polished for merge
by arch1t3cht and torque, but several more people
contributed bits and pieces. They are listed in the following.

Reviewed-by: Eli Schwartz <eschwartz93@gmail.com>
Reviewed-by: Oneric <oneric@oneric.stub>

Co-authored-by: arch1t3cht <arch1t3cht@gmail.com>
Co-authored-by: torque <torque@users.noreply.github.com>
Co-authored-by: Luni-4 <luni-4@hotmail.it>
Co-authored-by: Funami580 <Funami580@users.noreply.github.com>
Co-authored-by: woclass <5158738+inkydragon@users.noreply.github.com>
Co-authored-by: line0 <line0@inbox.lv>
Co-authored-by: Myaamori <myaamori1993@gmail.com>
Co-authored-by: Xavier Claessens <xavier.claessens@collabora.com>