Make STRVAL macros strdup-fail safe
Strictly speaking this is not neccessary in regular processing as the
event and font fields are intialised with zeroes anyway and there's no
old value to fallback too.
However when processing style overrides, this can prevent a non-null
value being replaced by NULL.

This commit also gets rid of the unneccessary '!= NULL' check. Passing
NULL to free is well defined and safe.