wrap_lines: do not merge linebreaks
Merging linebreaks will always result in text overflow its designated
bounds, but integer overflows or other bugs may make it seems like it
would be beneficial. If this is not prevented it may also lead to
out-of-bound reads as in CVE-2016-7969.

This replaces and reverts
https://github.com/libass/libass/commit/f4f48950788b91c6a30029cc28a240b834713ea7
which did not prevent linebreak-merges and turned
hard linebreaks into soft ones on merge.