Print libass version
This could help us in future bug reports and we already do so for
shapers and freetype. Also include a guess about the source code
version, which may not be entirely reliable. Eg if custom tarballs are
involved or if patches are added after configure was called, the value
will be wrong.
But hopefully it will be "good enough" for most cases.