directwrite: support WinRT/UWP, which forbid LoadLibrary
Normally, we delay loading Dwrite.dll until runtime to allow building
and running DirectWrite-enabled libass binaries on old Windows versions
that lack DirectWrite. However, this is forbidden in WinRT/UWP.

I don't think our configure script can be used to produce WinRT/UWP code.
But if someone is building us for WinRT/UWP (with a custom build setup),
our code will now work without requiring extra patches.
The builder is responsible for linking to Dwrite.dll at build time.
DirectWrite is present in all versions of Windows that support WinRT/UWP,
so we lose nothing by requiring it.

Older Windows SDKs (Microsoft and MinGW alike) lack <winapifamily.h>,
so include it only if we really need it.

Based on VLC patch:

https://code.videolan.org/videolan/vlc/-/commit/eedb57aa96d2bc0046a6da2e081c75ae9edf8fd5
https://github.com/videolan/vlc/commit/eedb57aa96d2bc0046a6da2e081c75ae9edf8fd5

Note: the VLC patch retained an unconditional call to FreeLibrary
in destroy_provider. However, FreeLibrary does not actually expect
NULL as argument, so this commit adds a guard to that call.
Perhaps FreeLibrary(NULL) simply fails cleanly and that's why
this has not caused VLC problems, but we should not rely on this.