coretext: log *all* attributes of each CTFontDescriptor
Core Text "helpfully" reduces the "attributes" dictionary to the smallest
set that supposedly uniquely identifies the font, which really means just
a lone name attribute, as names are supposed to be unique in Core Text.
But there are more attributes, and when font deduplication is disabled,
we can actually get multiple fonts with the same name but different paths!
which can't be distinguished looking at the reduced attributes dictionary.