Support Core Text on Mac OS X 10.5
Based on the logic in HarfBuzz's hb-coretext. See:
 1. https://github.com/harfbuzz/harfbuzz/pull/952
 2. https://github.com/harfbuzz/harfbuzz/pull/1180

Core Text was introduced in 10.5, so this is as far back as we can go.

Tested on ppc32 10.5.8 in:
https://github.com/libass/libass/issues/595#issuecomment-1324548120
Manrope Regular failed to be selected by full name,
but other fonts (including other faces of Manrope) worked fine,
so this should be good enough to avoid requiring Fontconfig.

I have tested that the code in the 10.5-compatible branch still works
on 10.13.6 (the newest macOS instance I have access to), but the API
involved is officially "deprecated" and produces deprecation warnings
during compilation unless the build's deployment target is set to a
version below 10.8. In addition, it's not unimaginable that it might
work worse in some specific situations now or in the future: for example,
it uses FSRef, which, according to the docs, is "designed to work
with 32-bit inode numbers", which "may result in performance issues"
on newer systems. As far as I'm aware, Apple doesn't tend to remove
APIs/ABIs completely except when combined with architecture changes,
but it seems the overall safest course of action is:
  * to prefer the newer API if it is available at runtime,
  * and to avoid referencing the older API at all
    if it's known at compile-time to be unnecessary.

To maximize build-environment compatibility, use AvailabilityMacros.h,
including MAC_OS_X_VERSION_MAX_ALLOWED, and a weak-import address check:

  * Recent versions of Apple's compiler support __builtin_available for
    OS version detection, and so does modern non-Apple Clang. This is what
    Apple recommends nowadays. However, none of Apple's official compilers
    on macOS up to 10.10 have had this builtin, and no third-party
    compilers besides Clang have ever had it. Non-Clang compilers may
    have issues with Apple's more recent SDK headers anyway, but they're
    perfectly viable with older SDK releases and all the more likely
    to be used when the build is being performed on an older machine,
    e. g. when targeting that same machine, which is precisely where
    the 10.5-compatible code is most likely to be relevant.
    On 10.5 or nearby versions, the build is most likely to use one of
    Apple's old compilers or a custom-built modern upstream GCC.

    Apple's older method of checking for availability at runtime
    is to check whether the symbol's address is NULL, which works in all
    of Apple's compilers and in upstream Clang and GCC, so use this.

    This produces a compiler warning in some situations: if the deployment
    target is 10.6+, the address is always non-NULL, so the condition is
    trivially true, which warrants a warning. For Clang and for GCC 12+,
    silence this warning by casting the address to an integer.

    (Note: Clang warns when __builtin_available is unused, but only
    for symbols introduced in macOS 10.13+, at which point other
    contemporary or modern compilers can no longer compile
    ass_coretext due to errors in system headers.)

  * We need _some_ way to determine whether the newer symbol is declared
    at all. There are multiple macros we could use, but in older
    versions of Apple's SDK, Core Text headers have reacted to the
    MAC_OS_X_VERSION_MAX_ALLOWED macro and marked newer APIs
    "unavailable" (making any use of them a compilation error)
    even if they were known to that SDK. Hence, use this same macro
    (which is still available in newer SDKs although it no longer marks
    specifically Core Text's symbols unavailable) as our detection
    mechanism.